+++ util/run.sh
@@ -522,6 +522,7 @@ while $RUN_MENU_AGAIN; do
                                       "4" "Read Serial Dynon Skyview" \
                                       "5" "Read Serial Garmin G3x" \
                                       "6" "Test Stratux and iLevil WiFi connection" \
+                                      "7" "I2C Test (Pi only)" \
                                       2>&1 1>&3)
                     exit_status=$?
                     exec 3>&-
@@ -534,6 +535,7 @@ while $RUN_MENU_AGAIN; do
                             4) FULL_COMMAND="python3 $TRONVIEW_DIR/util/tests/serial_read.py -s" ;;
                             5) FULL_COMMAND="python3 $TRONVIEW_DIR/util/tests/serial_read.py -g" ;;
                             6) FULL_COMMAND="$TRONVIEW_DIR/util/tests/test_stratux_wifi.sh" ;;
+                            7) FULL_COMMAND="python3 $TRONVIEW_DIR/util/tests/i2c_test.py" ;;
                         esac
                         echo "Running: $FULL_COMMAND"
                         #exit 0