+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - update from git. and reload run.sh
 - main.py fixes for 100 inputs
 - added i2c test script
-
+- bno055 and bno085 init fixes. auto set address for 2nd imu.
 
 
 ## [0.0.20] - 2024-12-06